Convert C-style (Foo *)x to &dynamic_cast<Foo &>(*x) in neocompiler.
Use static_cast for non-class casts.

On a hunt for memory corruption and a null pointer somewhere it
shouldn't be.  Haven't found it yet, but this make the code more
robust anyway by detecting bad casts before their results are used.

Many C-style casts still remain in the paleocompiler and simulator.